// mypay.sources/mypay4-fe/projects/mypay4-fe-operatore/src/app/model/ente.ts
/*
 *     MyPay - Payment portal of Regione Veneto.
 *     Copyright (C) 2022  Regione Veneto
 *
 *     This program is free software: you can redistribute it and/or modify
 *     it under the terms of the GNU Affero General Public License as
 *     published by the Free Software Foundation, either version 3 of the
 *     License, or (at your option) any later version.
 *
 *     This program is distributed in the hope that it will be useful,
 *     but WITHOUT ANY WARRANTY; without even the implied warranty of
 *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 *     GNU Affero General Public License for more details.
 *
 *     You should have received a copy of the GNU Affero General Public License
 *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
import { DateTime } from 'luxon';
import {
    Ente as CommonEnte, MapperDef, MapperType
} from 'projects/mypay4-fe-common/src/public-api';

import { AnagraficaStato } from './anagrafica-stato';

export class Ente extends CommonEnte {
  public static readonly MAPPER_DEF = [
    new MapperDef(MapperType.DateTime,'dtUltimaModifica','local-date-time'),
    new MapperDef(MapperType.DateTime,'dtAvvio','local-date'),
  ];

  mygovEnteId: number;
  codIpaEnte: string;
  codiceFiscaleEnte: string;
  deNomeEnte: string;
  emailAmministratore: string;
  dtUltimaModifica: DateTime;
  codRpDatiVersDatiSingVersIbanAccredito: string;
  enteSilInviaRispostaPagamentoUrl: string;
  dePassword: string;
  paaSILInviaCarrelloDovutiHash: string;
  codRpDatiVersDatiSingVersBicAccreditoSeller: boolean;
  deRpEnteBenefDenominazioneBeneficiario: string;
  deRpEnteBenefIndirizzoBeneficiario: string;
  deRpEnteBenefCivicoBeneficiario: string;
  codRpEnteBenefCapBeneficiario: string;
  deRpEnteBenefLocalitaBeneficiario: string;
  deRpEnteBenefProvinciaBeneficiario: string;
  codRpEnteBenefNazioneBeneficiario: string;
  deRpEnteBenefTelefonoBeneficiario: string;
  deRpEnteBenefSitoWebBeneficiario: string;
  deRpEnteBenefEmailBeneficiario: string;
  applicationCode: string;
  codCodiceInterbancarioCbill: string;
  deInformazioniEnte: string;
  cdStatoEnte: AnagraficaStato;
  deAutorizzazione: string;
  codTipoEnte: string;
  dtAvvio: DateTime;
  linguaAggiuntiva: string;
  flagInsertDefaultSet: boolean;
}